+++ lib/pah/partials/_database.rb
@@ -1,5 +1,5 @@
 copy_static_file 'config/database.yml'
-gsub_file 'config/database.yml', /PROJECT/, @app_name
+gsub_file 'config/database.yml', /PROJECT/, app_name
 git add: 'config/database.yml'
 git_commit 'Adding config/database.yml.'