+++ lib/pah/partials/_readme.rb
@@ -1,7 +1,7 @@
 puts "Adding default README... ".magenta
 
 copy_static_file 'README.md'
-gsub_file 'README.md', /PROJECT/, @app_name
+gsub_file 'README.md', /PROJECT/, app_name
 
 git add: 'README.md'
 git_commit 'Adding README file.'